Added (starts-with) and fixed bugs with parser
(starts-with) allows you to test if a string starts with another string.
This was added to allow for more string testing, and it returns either
true or false.  It is not fully implemented as of yet and WILL cause
warnings, I will hopefully get full error checking implemented and
pushed soon as time permits.  Do not expect to use this function with
undefined behavior or strange things may happen.

Otherwise a bug was fixed involving the function parser, which meant
that local variables would be considered as undefined when passed to
functions.  As well a bug with parsing comments was fixed, where placing
a comment would make an error about a variable being undefined.  This
was fixed by making sure comments didn't then invoke the executor.
Please open an issue if either of these changes break any existing code.