Improved CMakeLists.txt, fixed bugs, closes #6
CMakeLists.txt should work much better now on windows, and also now
allows the building of crow as a shared library or static library
depending on what value is passed.  -DLIBCROW=STATIC for a static
library, and -DLIBCROW=SHARED for a shared library.  As well, now the
version string is recorded, which should make using specific versions of
libcrow much easier.

The bugs fixed were all minor.  Mostly just typos or semantics.  One
memory free (which may or may not be allowed, I'm not sure, it seems
obvious enough, but there was some reason I didn't do it earlier).  The
biggest issue fixed was #6 which was a NULL terminator problem.