Modified how functions worked and updated CRO_eval
Functions now accept arguements using spaces. For some reason it never
occured to me that commas are perfectly okay to have in variable names,
so i was using them as separators. However, with spaces, they are alreay
separators in the langauge, so lets use them instead. Also added the
subroutine function (=>) pending a large rewrite of how (if) and other
conditional statements work. The subroutine function is basically just a
function that takes zero arguements, so you don't need to specify any.

Other than that, for some reason CRO_eval was still using an extremely
old implementation of interpreting, so I updated it to feature the more
up to date and robust model that CRO_evalFile as well as the REPL are
using.