Added exposable type system
Now types are able to be programatically exposed to crow. For example if
a library is added to provide an Image type, all you would need to do is
add a CRO_exposeType(printImage) where printImage is a print function
able to display the type in some visual way. This allows more
flexibility in 3rd party libraries. CRO_None is also now referred to as
CRO_Undefined to not conflict with the macro CRO_NONE for exit codes.