[#28] Format all files using regex.
core.c was done manually, however every other file was done with regex, so there is the possibility that some formatting is still incorrect and was not caught by the regex search and replace, or that some formatting was done wrong. The codebase still builds so build-ability was not impacted. For that reason however, #28 will remain open until I feel like I can comfortably say all files are up to the formatting specification.