[#29] Rewrite variables to use closures
How did this not take days to implement. Crow uses closures now, which means lambdas work now! Closes #29 and also closes #26. Since the closures take the place of the variable list, #25 is no longer relevant to the code base, however structs will be rewritten to use trees in the near future, so keeping open for that.