[#23] Rewrite garbage collector
This was... something i guess. So basically Crow now has a very robust type system that I may eventually completely gut because it may be more than we need. The garbage collector now works for all types, even for CRO_Files which were broken for a while. CRO_File also now just allocates to allocations, pending a removal of the FileDescriptor system to make it more streamlined. This new garbage collector has been tested, but errors are sure to crop up. Should be considered unstable for now, feel free to disable it using the flag. This closes #36 and closes #38.