[#39] Fix issues with the garbage collector
This introduces a new flag for the allocations which basically says "skip checking this, it's in use right now" which any function can toggle on when it is using an array, so that even if the data is no longer accessible by a variable, the memory isn't free'd. This also removes a lot of dead code for dealing with file, since those were reworked after the garbage collector update. Closes #39