[#48][#52] Implement (local) and fix string bug.
Implemented (local), but half way through doing that, found and fixed a bug in the parser dealing with an uninitialized value causing strings to not terminate when they were supposed to. Also fixed an issue with CRO_While missing a bracket which may cause issues on some platforms (but I think clang was protecting me from it). Closes #48 and closes #52.