[#55] Implement new Garbage Collection system
Three days worth of work. THREE DAYS! Well, finally it's done and its _mostly_ working, and a hell of a lot better than the last GC. Im going to keep #55 open until all the bugs are ironed out of the current implementation, because believe me, there are a TON of bugs still in it. Also added tests as a way of testing how complete the new GC was, and they're pretty useful to make sure the language doesn't see any regressions. Regressions are probably going to be common after this push, so please tag regressions with the "Regression" tag in issues.