[No Issue] Add faster variable lookup
This destructive change is behind a flag, but will greatly speed up
variable lookups. This instead of storing variable names as symbols,
stores them as hashes, which are much faster to compare against. This
is, however, not how Lisps typically work, so it is behind a flag for
anyone who wants to use it.